requirements.txt
Django==5.0.1
django-extensions==4.1
Pillow==11.3.0
python-docx==1.2.0
PyMuPDF==1.26.4
langchain==0.3.27
langchain-openai==0.3.35
langchain-community==0.3.31
openai==1.109.1
faiss-cpu==1.12.0
chromadb==1.1.1
whitenoise==6.11.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
gunicorn==23.0.0
dj-database-url==2.1.0
